+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Add `RSpec/Dialect` enforces custom RSpec dialects. ([@gsamokovarov][])
 * Fix redundant blank lines in `RSpec/MultipleSubjects`'s autocorrect. ([@pirj][])
 * Drop support for ruby `2.2`. ([@bquorning][])
+* Remove `AggregateFailuresByDefault` config option of `RSpec/MultipleExpectations`. ([@pirj][])
 
 ## 1.32.0 (2019-01-27)